Fix -Wall -Werror compliance for all test targets (--enable-all)
Source fixes:
- fts5Int.hpp/fts5_buffer.cpp: const-qualify sqlite3Fts5BufferAppendPrintf fmt param
- fts5_main.cpp: const_cast for empty string fallback
- fts5_expr.cpp: fix sqlite3Fts5ParserTrace forward decl to const char*
- fts3_tokenizer.cpp: const-qualify queryTokenizer/registerTokenizer params,
  mark rc as [[maybe_unused]]
- fts3_test.cpp: const-qualify zName in TestnmSubcmd struct
- fts5_tcl.cpp: const-qualify zName in Cmd struct
- geopoly.cpp: const_cast for idxStr string literal assignments
- sqlite3session.cpp: add static_casts for sessionMalloc64 returns
- test_session.cpp: add static_casts for Tcl_NewByteArrayObj void* args,
  mark assert-only rc variables as [[maybe_unused]], cast ckalloc
- analyze.cpp: fix reinterpret_cast to const, const-qualify decodeIntArray
  and zIndex, fix static_cast Table* to reinterpret_cast
- bitvec.cpp: use unsigned loop var in showBitvec (sign-compare)
- btree.cpp: cast szScratch assert to u64 (sign-compare)
- main.cpp: remove volatile from TESTCTRL_ASSERT variable (C++20)
- mutex_unix.cpp: replace volatile ++/-- with explicit assignment (C++20)
- sqliteInt.hpp: const-qualify sqlite3ParserTrace declaration
- treeview.cpp: const_cast for "(null)" fallback string
- util.cpp: cast sizeof to int in decimal assertions (sign-compare)
- test_btree.cpp: const-qualify zMode
- test_demovfs.cpp: cast iAmt to size_t for comparison
- shell.cpp.in: const-qualify azColorNames array
- lempar.c: const-qualify yyTracePrompt and ParseTrace parameter

Test file renames (.c -> .cpp) and C++ porting:
- test/fuzzcheck: rename, fix includes (.h -> .hpp), add void* casts,
  const-qualify string arrays, fix extern init signatures, fix sign-compare
- test/fuzzinvariants: rename, fix includes, add void* casts, const arrays
- test/ossfuzz: rename, fix includes
- test/sessionfuzz: rename, fix #include sqlite3.cpp, add void* casts,
  add OMIT_ZLIB guard
- test/vt02: rename, fix includes, add void* casts
- main.mk: update all references to renamed test files

977,176 tests pass (mdevtest). Remaining 3 failures are pre-existing:
vtabH.test (memory leak), shell1.test (expects .c filenames in output).

+++ src/btree.cpp
@@ -8381,7 +8381,7 @@ static int balance_nonroot(
      + nMaxCells*sizeof(u16)                       /* b.szCell */
      + pBt->pageSize;                              /* aSpace1 */
 
-  assert( szScratch<=7*(int)pBt->pageSize );
+  assert( szScratch<=(u64)7*pBt->pageSize );
   b.apCell = static_cast<u8**>(sqlite3StackAllocRaw(0, szScratch ));
   if( b.apCell==0 ){
     rc = SQLITE_NOMEM_BKPT;
